Move register/unregister inside DNS callback
