+++ app-tracking-protection/vpn-impl/src/main/java/com/duckduckgo/mobile/android/vpn/pixels/DeviceShieldPixels.kt
@@ -349,6 +349,8 @@ interface DeviceShieldPixels {
 
     fun reportVpnSnoozedStarted()
     fun reportVpnSnoozedEnded()
+
+    fun reportMotoGFix()
 }
 
 @ContributesBinding(AppScope::class)
@@ -745,6 +747,10 @@ class RealDeviceShieldPixels @Inject constructor(
     override fun reportVpnSnoozedEnded() {
         tryToFireDailyPixel(DeviceShieldPixelNames.VPN_SNOOZE_ENDED_DAILY)
         firePixel(DeviceShieldPixelNames.VPN_SNOOZE_ENDED)
+    }
+
+    override fun reportMotoGFix() {
+        tryToFireDailyPixel(DeviceShieldPixelNames.VPN_MOTO_G_FIX_DAILY)
     }
 
     private fun suddenKill() {
